+++ doc/source/conf.py
@@ -269,5 +269,8 @@ texinfo_documents = [
 # If true, do not generate a @detailmenu in the "Top" node's menu.
 #texinfo_no_detailmenu = False
 
+# for complete documentation of API
+autodoc_default_options = {"members": True}
+
 # autosummary settings
 autosummary_generate = True